Refactor - ResultView 기능에 따른 클래스 분리
